requirements.txt
opencv-python
numpy
pandas
streamlit
Pillow
